requirements.txt
lldb-python
setuptools
